URL encoding.  Refactoring of PrivateMusicCataloguerCLI.
Moved all but main(String...) from PrivateMusicCataloguerCLI to a new class named PrivateMusicCataloguer.  Added logic to PrivateMusicCataloguer for encoding the song and album art URLs.
