+++ src/main/resources/META-INF/persistence.xml
@@ -23,7 +23,8 @@
         <class>f18a14c09s.integration.alexa.music.data.ArtSource</class>
         <properties>
             <!--<property name="hibernate.show_sql" value="true"/>-->
-            <property name="hibernate.hbm2ddl.auto" value="create-drop"/>
+            <!--<property name="hibernate.hbm2ddl.auto" value="create-drop"/>-->
+            <property name="hibernate.hbm2ddl.auto" value="create"/>
         </properties>
     </persistence-unit>
 </persistence>
